refactor(core): extract instance module from location services
